Rename as defaultTopicTriggers-get middleware
