+++ app/routes/defaultTopicTriggers/index.js
@@ -2,9 +2,9 @@
 
 const express = require('express');
 
-const defaultTopicTriggersMiddleware = require('../../../lib/middleware/defaultTopicTriggers/index');
+const defaultTopicTriggersMiddleware = require('../../../lib/middleware/defaultTopicTriggers/index/defaultTopicTriggers-get');
 
-const router = express.Router({ mergeParams: true });
+const router = express.Router();
 
 router.use(defaultTopicTriggersMiddleware());
