+++ app/routes/topics/index.js
@@ -4,7 +4,7 @@ const express = require('express');
 
 const getTopicsMiddleware = require('../../../lib/middleware/topics/index/topics-get');
 
-const router = express.Router({ mergeParams: true });
+const router = express.Router();
 
 router.use(getTopicsMiddleware());
